backend/requirements.txt
python-dotenv
google-generativeai
SpeechRecognition
pyaudio
requests
fastapi
uvicorn
python-multipart

